refactor: extract composite pattern
